Move to details nfc package
