+++ core-lib/src/main/java/org/idpass/smartscanner/lib/nfc/details/AdditionalPersonDetails.java
@@ -15,7 +15,7 @@
  *
  *
  */
-package org.idpass.smartscanner.lib.nfc;
+package org.idpass.smartscanner.lib.nfc.details;
 
 import java.util.List;
